Fix AttributeError raise on ConnectionError
The 'mkdirs' method was raising an AttributeError when the connection
with the server was unsuccessful. This was due to a try-except that was
handling all exception classes and tried accessed to access to a
the 'actual_code' property which is not defined for the Exception class.
The approach to fix the issue was to stop catching all exceptions and
just let them be passed out of the 'mkdirs' method.